Move database-helpers into model-helpers
